+++ src/assets/styles/display.ts
@@ -1,10 +1,10 @@
 import { formatBreakpoint } from '$lib/utils/format-class.js';
 
-export function positionClassName(screen: string, values: Array<string>) {
+export function displayClassName(screen: string, values: Array<string>) {
 	let css = '';
 	for (const element of values) {
 		css += `${formatBreakpoint(screen)}${element} {\n`;
-		css += `position: ${element};\n`;
+		css += `display: ${element};\n`;
 		css += `}\n`;
 	}
 	return css;